requirements.txt
PyQt5>=5.10,<5.14.1
PyQtGraph
PySerial
PyYaml
